# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API's Xlearn",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://servicios.10xconsultores.org"
        },
        {
            "url": "http://127.0.0.1:8000"
        },
        {
            "url": "https://servicios.asstiseguridadsocial.com"
        }
    ],
    "paths": {
...
        "/api/v1/course/list": {
            "get": {
                "tags": [
                    "Courses"
                ],
                "summary": "Mostrar cursos",
                "operationId": "02708c6cb845c5a82a34d5daf197c9a9",
                "parameters": [
                    {
                        "name": "offset",
                        "in": "query",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mostrar todos los cursos.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "response": {
                                            "hc:length": 0,
                                            "hc:total": 0,
                                            "hc:offset": 0,
                                            "hc:limit": 0,
                                            "hc:next": "next page end-point ",
                                            "hc:previous": "previous page end-point ",
                                            "_rel": "course",
                                            "_embedded": {
                                                "course": [
                                                    {
                                                        "id": 0,
                                                        "name": "",
                                                        "description": "",
                                                        "state": 0,
                                                        "free_video": "",
                                                        "file_path": "",
                                                        "video_path": "",
                                                        "area_id": 0,
                                                        "programs_id": 0,
                                                        "created_at": "2022-06-11T23:21:42.000000Z",
                                                        "updated_at": "2022-06-12T00:46:06.000000Z",
                                                        "areas": {
                                                            "id": 0,
                                                            "name": "",
                                                            "description": "",
                                                            "file_path": "",
                                                            "state": 0,
                                                            "created_at": "2022-06-11T23:21:42.000000Z",
                                                            "updated_at": "2022-06-12T00:46:06.000000Z"
                                                        }
                                                    }
                                                ]
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "message": "Mensaje de error"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/course/show_area/{areaId}": {
            "get": {
                "tags": [
                    "Courses"
                ],
                "summary": "Mostrar curso por areas",
                "operationId": "bef255018c2e979d2db5045f9915def8",
                "parameters": [
                    {
                        "name": "offset",
                        "in": "query",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "areaId",
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mostrar todos los cursos del area.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "response": {
                                            "hc:length": 0,
                                            "hc:total": 0,
                                            "hc:offset": 0,
                                            "hc:limit": 0,
                                            "hc:next": "next page end-point ",
                                            "hc:previous": "previous page end-point ",
                                            "_rel": "course",
                                            "_embedded": {
                                                "course": [
                                                    {
                                                        "id": 0,
                                                        "name": "",
                                                        "description": "",
                                                        "state": 0,
                                                        "free_video": "",
                                                        "file_path": "",
                                                        "video_path": "",
                                                        "area_id": 0,
                                                        "programs_id": 0,
                                                        "created_at": "2022-06-11T23:21:42.000000Z",
                                                        "updated_at": "2022-06-12T00:46:06.000000Z",
                                                        "areas": {
                                                            "id": 0,
                                                            "name": "",
                                                            "description": "",
                                                            "file_path": "",
                                                            "state": 0,
                                                            "created_at": "2022-06-11T23:21:42.000000Z",
                                                            "updated_at": "2022-06-11T23:21:42.000000Z"
                                                        }
                                                    }
                                                ]
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "message": "Mensaje de error"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/course/show_user/{userId}": {
            "get": {
                "tags": [
                    "Courses"
                ],
                "summary": "Mostrar curso por usuario",
                "operationId": "fc67351613a36af84c2bbf2c54274ba6",
                "parameters": [
                    {
                        "name": "offset",
                        "in": "query",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "userId",
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mostrar todos los cursos del usuario.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "response": {
                                            "hc:length": 0,
                                            "hc:total": 0,
                                            "hc:offset": 0,
                                            "hc:limit": 0,
                                            "hc:next": "next page end-point ",
                                            "hc:previous": "previous page end-point ",
                                            "_rel": "courses",
                                            "_embedded": {
                                                "courses": [
                                                    {
                                                        "id": 0,
                                                        "name": "",
                                                        "description": "",
                                                        "state": 0,
                                                        "free_video": "",
                                                        "file_path": "",
                                                        "video_path": "",
                                                        "area_id": 0,
                                                        "programs_id": 0,
                                                        "created_at": "2022-06-11T23:21:42.000000Z",
                                                        "updated_at": "2022-06-12T00:46:06.000000Z"
                                                    }
                                                ]
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "message": "Mensaje de error"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}